[#1] Fix : 폴더구조 변경
