--- src/main/java/Journey/Together/domain/SocialType.java
@@ -1,11 +1,0 @@
-package Journey.Together.domain;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum SocialType {
-    KAKAO,
-    NAVER;
-}
