--- src/main/java/Journey/Together/exception/model/BadRequestException.java
@@ -1,9 +1,0 @@
-package Journey.Together.exception.model;
-
-import Journey.Together.exception.Error;
-public class BadRequestException extends CustomException{
-	public BadRequestException(Error error, String message) {
-		super(error, message);
-	}
-
-}
