--- src/main/java/Journey/Together/exception/model/UnauthorizedException.java
@@ -1,9 +1,0 @@
-package Journey.Together.exception.model;
-
-import Journey.Together.exception.Error;
-public class UnauthorizedException extends CustomException{
-	public UnauthorizedException(Error error, String message) {
-		super(error, message);
-	}
-
-}
