--- src/main/java/Journey/Together/exception/model/UnprocessableEntityException.java
@@ -1,9 +1,0 @@
-package Journey.Together.exception.model;
-
-import Journey.Together.exception.Error;
-
-public class UnprocessableEntityException extends CustomException{
-	public UnprocessableEntityException(Error error, String message) {
-		super(error, message);
-	}
-}
