--- src/main/java/Journey/Together/external/aws/ImagePresignedUrlResponse.java
@@ -1,7 +1,0 @@
-package Journey.Together.external.aws;
-
-public record ImagePresignedUrlResponse(String fileName, String preSignedUrl) {
-	public static ImagePresignedUrlResponse of(String fileName, String url){
-		return new ImagePresignedUrlResponse(fileName, url);
-	}
-}
