--- src/main/java/Journey/Together/external/aws/PresignedUrlVO.java
@@ -1,7 +1,0 @@
-package Journey.Together.external.aws;
-
-public record PresignedUrlVO(String fileName, String url) {
-	public static PresignedUrlVO of(String fileName, String url){
-		return new PresignedUrlVO(fileName, url);
-	}
-}
