--- src/main/java/Journey/Together/external/aws/S3Service.java
@@ -1,57 +1,0 @@
-package Journey.Together.external.aws;
-
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-import org.springframework.web.multipart.MultipartFile;
-import software.amazon.awssdk.core.sync.RequestBody;
-import software.amazon.awssdk.services.s3.S3Client;
-import software.amazon.awssdk.services.s3.model.DeleteObjectRequest;
-import software.amazon.awssdk.services.s3.model.PutObjectRequest;
-
-import java.io.IOException;
-import java.util.UUID;
-
-@Component
-public class S3Service {
-
-    private final String bucketName;
-    private final AWSConfig awsConfig;
-
-    public S3Service(@Value("${aws-property.s3-bucket-name}") final String bucketName, AWSConfig awsConfig) {
-        this.bucketName = bucketName;
-        this.awsConfig = awsConfig;
-    }
-
-
-    public String uploadImage(String directoryPath, MultipartFile image) throws IOException {
-        final String key = directoryPath + generateImageFileName();
-        final S3Client s3Client = awsConfig.getS3Client();
-
-        PutObjectRequest request = PutObjectRequest.builder()
-                .bucket(bucketName)
-                .key(key)
-                .contentType(image.getContentType())
-                .contentDisposition("inline")
-                .build();
-
-        RequestBody requestBody = RequestBody.fromBytes(image.getBytes());
-        s3Client.putObject(request, requestBody);
-        return key;
-    }
-
-    public void deleteImage(String key) throws IOException {
-        final S3Client s3Client = awsConfig.getS3Client();
-
-        s3Client.deleteObject((DeleteObjectRequest.Builder builder) ->
-                builder.bucket(bucketName)
-                        .key(key)
-                        .build()
-        );
-    }
-
-
-    private String generateImageFileName() {
-        return UUID.randomUUID().toString() + ".jpg";
-    }
-
-}
