+++ src/main/java/Journey/Together/global/exception/Success.java
@@ -1,4 +1,4 @@
-package Journey.Together.exception;
+package Journey.Together.global.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
