+++ src/main/java/Journey/Together/global/exception/model/BadRequestException.java
@@ -1,0 +1,10 @@
+package Journey.Together.global.exception.model;
+
+import Journey.Together.global.exception.Error;
+
+public class BadRequestException extends CustomException {
+	public BadRequestException(Error error, String message) {
+		super(error, message);
+	}
+
+}
