+++ src/main/java/Journey/Together/global/exception/model/ForbiddenException.java
@@ -1,0 +1,10 @@
+package Journey.Together.global.exception.model;
+
+import Journey.Together.global.exception.Error;
+
+public class ForbiddenException extends CustomException {
+	public ForbiddenException(Error error, String message) {
+		super(error, message);
+	}
+
+}
