+++ src/main/java/Journey/Together/global/exception/model/NotFoundException.java
@@ -1,6 +1,7 @@
-package Journey.Together.exception.model;
+package Journey.Together.global.exception.model;
 
-import Journey.Together.exception.Error;
+import Journey.Together.global.exception.Error;
+
 public class NotFoundException extends CustomException {
 	public NotFoundException(Error error, String message) {
 		super(error, message);
