+++ src/main/java/Journey/Together/global/exception/model/UnauthorizedException.java
@@ -1,0 +1,10 @@
+package Journey.Together.global.exception.model;
+
+import Journey.Together.global.exception.Error;
+
+public class UnauthorizedException extends CustomException {
+	public UnauthorizedException(Error error, String message) {
+		super(error, message);
+	}
+
+}
