+++ src/main/java/Journey/Together/global/exception/model/net/UnknownHostException.java
@@ -1,5 +1,7 @@
-package Journey.Together.exception.model.net;
-import Journey.Together.exception.Error;
+package Journey.Together.global.exception.model.net;
+
+import Journey.Together.global.exception.Error;
+
 public class UnknownHostException extends CustomJavaNetException {
 	public UnknownHostException(Error error, String message) {
 		super(error, message);
